move Navigation to common/Header
